backend/recipe_app/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.7
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
